Make the library a single header file
