--- src/custom_types.h
@@ -1,7 +1,0 @@
-#ifndef GCP_CUSTOM_TYPES_H
-#define GCP_CUSTOM_TYPES_H
-
-using GenotypeConfidence = double;
-using GenotypePercentile = double;
-
-#endif // GCP_CUSTOM_TYPES_H
